Moved functions from OrganismBase to OrganismTypeBase.
