+++ apps/MABE/core/OrganismBase.h
@@ -21,24 +21,10 @@
 namespace mabe {
 
   class OrganismBase {
-  protected:
-    struct FunctionBase { template <typename T> auto Convert(); };
-    template <typename T> struct Function : public FunctionBase {
-      std::function<T> fun;
-      Function(std::function<T> _fun) : fun(_fun) { ; }
-    };
-
-    using fun_ptr_t = emp::Ptr<FunctionBase>;
-    using fun_vec_t = emp::vector<fun_ptr_t>;
-    std::map< std::string, fun_vec_t > fun_map;
   public:
     OrganismBase() { ; }
     virtual ~OrganismBase() { ; }
   };
-
-  template <typename T> auto OrganismBase::FunctionBase::Convert() {
-    return (emp::Ptr<OrganismBase::Function<T>>) dynamic_cast< OrganismBase::Function<T> >(this);
-  }
 
 }
